+++ Controllers/Business/ProcessoController.cs
@@ -1,8 +1,12 @@
-﻿using Calcular.CoreApi.Models;
+﻿using Calcular.CoreApi.Common;
+using Calcular.CoreApi.Models;
 using Calcular.CoreApi.Models.Business;
+using Calcular.CoreApi.Models.ViewModels;
 using Calcular.CoreApi.Shared;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -13,44 +17,158 @@ namespace Calcular.CoreApi.Controllers.Business
     public class ProcessoController : Controller
     {
         private readonly ApplicationDbContext db;
+        private readonly UserManager<User> userManager;
 
-        public ProcessoController(ApplicationDbContext db)
+        public ProcessoController(UserManager<User> userManager, ApplicationDbContext db)
         {
             this.db = db;
+            this.userManager = userManager;
         }
 
         [HttpGet]
         public IActionResult GetAll([FromQuery] string filter)
         {
-            var result = db.Processos.Where(x => string.IsNullOrEmpty(filter)
-                                             || x.Numero.Contains(filter)
-                                             || x.Reu.Contains(filter)
-                                             || x.Autor.Contains(filter)
-                                             || x.Advogado.Nome.Contains(filter));
+            var query = db.Processos
+                            .Include(x => x.Advogado)
+                            .Include(x => x.Honorarios)
+                            .Where(x => string.IsNullOrEmpty(filter)
+                                        || x.Numero.Contains(filter)
+                                        || x.Reu.Contains(filter)
+                                        || x.Autor.Contains(filter))
+                            .OrderBy(x => x.Id)
+                            .Select(x => new ProcessoViewModel
+                            {
+                                Numero = x.Numero,
+                                Id = x.Id,
+                                Autor = x.Autor,
+                                Reu = x.Reu,
+                                Honorario = x.Honorario,
+                                Prazo = x.Prazo,
+                                Total = x.Total,
+                                Vara = x.Vara,
+                                Advogado = new Cliente
+                                {
+                                    Id = x.Advogado.Id,
+                                    Nome = x.Advogado.Nome
+                                },
+                            }).ToList();
 
-            return Ok(result.ToList());
+            return Ok(query);
+        }
+
+        [HttpGet]
+        [Route("numero")]
+        public IActionResult GetByNumber([FromQuery] string filter)
+        {
+            if (filter.Length < 3)
+                return BadRequest("Filtro deve conter no mínimo 3 caracteres.");
+
+            var result = db.Processos
+                            .Include(x => x.Advogado)
+                            .Include(x => x.Honorarios)
+                            .Include(x => x.ProcessoDetalhes).ThenInclude(x => x.User)
+                            .Where(x => string.IsNullOrEmpty(filter)
+                                             || x.Numero.Contains(filter))
+                            .Select(x => new ProcessoViewModel
+                            {
+                                Numero = x.Numero,
+                                Id = x.Id,
+                                Autor = x.Autor,
+                                Reu = x.Reu,
+                                Honorario = x.Honorario,
+                                Prazo = x.Prazo,
+                                Total = x.Total,
+                                Vara = x.Vara,
+                                Perito = x.Perito,
+                                Indicacao = x.Indicacao,
+                                Advogado = new Cliente
+                                {
+                                    Id = x.Advogado.Id,
+                                    Nome = x.Advogado.Nome
+                                },
+                                Honorarios = x.Honorarios.Select(h => new Honorario
+                                {
+                                    Id = h.Id,
+                                    Data = h.Data,
+                                    Registro = h.Registro,
+                                    Valor = h.Valor,
+                                    Prazo = h.Prazo,
+                                    Observacao = h.Observacao,
+                                    TipoPagamento = h.TipoPagamento,
+                                    NotaFiscal = h.NotaFiscal,
+                                }).ToList(),
+                                ProcessoDetalhes = x.ProcessoDetalhes.Select(p => new ProcessoDetalhe
+                                {
+                                    Id = p.Id,
+                                    Descricao = p.Descricao,
+                                    Data = p.Data,
+                                    User = new User
+                                    {
+                                        Id = p.User.Id,
+                                        Name = p.User.Name
+                                    }
+                                }).ToList()
+                            })
+                            .ToList();
+
+            return Ok(result);
         }
 
         [HttpGet("{id}")]
         public IActionResult GetById(int id)
         {
-            var result = db.Processos.Single(x => x.Id == id);
+            var result = db.Processos
+                            .Include(x => x.Advogado)
+                            .Include(x => x.Honorarios)
+                            .Include(x => x.ProcessoDetalhes).ThenInclude(x => x.User)
+                            .SingleOrDefault(x => x.Id == id);
+
             return Ok(result);
         }
 
         [HttpPost]
-        public IActionResult PostProcesso([FromBody] Processo Processo)
+        public IActionResult PostProcesso([FromBody] Processo processo)
         {
             try
             {
-                db.Processos.Add(Processo);
+                db.Processos.Add(processo);
+
+                if (processo.Local == TipoJusticaEnum.Outro && string.IsNullOrEmpty(processo.Numero))
+                    processo.Numero = processo.Id.ToString();
+
                 db.SaveChanges();
             }
             catch (Exception ex)
             {
                 Console.WriteLine(ex.Message);
             }
-            return Ok(Processo);
+            return Ok(processo);
+        }
+
+        [HttpPost]
+        [Route("detalhe")]
+        public IActionResult PostProcessoDetalhe([FromBody] ProcessoDetalhe processoDetalhe)
+        {
+            try
+            {
+                var user = userManager.GetUserAsync(HttpContext.User).Result;
+                processoDetalhe.UserId = user.Id;
+                processoDetalhe.Data = DateTime.Now;
+
+                var processo = db.Processos
+                                .Include(x => x.ProcessoDetalhes)
+                                .Single(x => x.Id == processoDetalhe.ProcessoId);
+
+                processo.ProcessoDetalhes.Add(processoDetalhe);
+                db.SaveChanges();
+
+                return Ok();
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                return BadRequest(ex.Message);
+            }
         }
 
         [HttpPut]
@@ -63,9 +181,10 @@ namespace Calcular.CoreApi.Controllers.Business
             item.Reu = newItem.Reu;
             item.Local = newItem.Local;
             item.Parte = newItem.Parte;
+            item.Vara = newItem.Vara;
             item.AdvogadoId = newItem.AdvogadoId;
-            item.IndicacaoId = newItem.IndicacaoId;
-            item.Honorario = newItem.Honorario;
+            item.Indicacao = newItem.Indicacao;
+            item.Perito = newItem.Perito;
 
             db.SaveChanges();
             return Ok(item);
@@ -74,19 +193,30 @@ namespace Calcular.CoreApi.Controllers.Business
         [HttpDelete("{id}")]
         public IActionResult DeleteProcesso(int id)
         {
-            var item = db.Processos.Single(x => x.Id == id);
-            db.Processos.Remove(item);
-            db.SaveChanges();
-            return Ok(item);
+            var item = db.Processos
+                        .Include(x => x.Honorarios)
+                        .Include(x => x.Servicos)
+                        .Single(x => x.Id == id);
+
+            if (item.Honorarios?.Count > 0 || item.Servicos?.Count > 0)
+            {
+                return BadRequest("O processo possúi Honorários/Serviços associados e não pode ser excluído");
+            }
+            else
+            {
+                db.Processos.Remove(item);
+                db.SaveChanges();
+                return Ok(item);
+            }
         }
 
         [Route("local")]
         [HttpGet]
         public IActionResult GetLocal()
         {
-            var result = Enum.GetValues(typeof(LocalEnum))
-                            .Cast<LocalEnum>()
-                            .Select(x => new KeyValuePair<int, string>((int)x, EnumHelpers.GetEnumDescription(x)));
+            var result = Enum.GetValues(typeof(TipoJusticaEnum))
+                            .Cast<TipoJusticaEnum>()
+                            .Select(x => new { Key = (int)x, Value = EnumHelpers.GetEnumDescription(x), Mask = EnumHelpers.GetAttributeOfType<MaskAttribute>(x) });
             return Ok(result);
         }
 
@@ -97,7 +227,46 @@ namespace Calcular.CoreApi.Controllers.Business
             var result = Enum.GetValues(typeof(ParteEnum))
                             .Cast<ParteEnum>()
                             .Select(x => new KeyValuePair<int, string>((int)x, EnumHelpers.GetEnumDescription(x)));
-            return Ok(result);  
+            return Ok(result);
+        }
+
+        [HttpGet("vara/{filter?}")]
+        public IActionResult GetVara([FromQuery] string filter = "")
+        {
+            var result = db.Processos.Where(x => string.IsNullOrEmpty(filter)
+                                             || x.Vara.ContainsIgnoreNonSpacing(filter))
+                                    .Select(x => x.Vara)
+                                    .Distinct()
+                                    .OrderBy(x => x)
+                                    .ToList();
+
+            return Ok(result);
+        }
+
+        [HttpGet("indicacao/{filter?}")]
+        public IActionResult GetIndicacao([FromQuery] string filter = "")
+        {
+            var result = db.Processos.Where(x => string.IsNullOrEmpty(filter)
+                                             || x.Indicacao.ContainsIgnoreNonSpacing(filter))
+                                    .Select(x => x.Indicacao)
+                                    .Distinct()
+                                    .OrderBy(x => x)
+                                    .ToList();
+
+            return Ok(result);
+        }
+
+        [HttpGet("perito/{filter?}")]
+        public IActionResult GetPerito([FromQuery] string filter = "")
+        {
+            var result = db.Processos.Where(x => string.IsNullOrEmpty(filter)
+                                             || x.Perito.ContainsIgnoreNonSpacing(filter))
+                                    .Select(x => x.Perito)
+                                    .Distinct()
+                                    .OrderBy(x => x)
+                                    .ToList();
+
+            return Ok(result);
         }
     }
 }
